+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 * `TabContainer.switcherPosition` has been moved to `TabContainerModel`. Please note that changes
   to `switcherPosition` are not supported on mobile, where the switcher will always appear
   beneath the container.
+* Mobile `Page` has changed - `Pages` are now wrappers around `Panels` that are designed to be used
+  with a `NavigationModel` or `TabContainer`. `Page` accepts the same props as `Panel`, meaning
+  uses of `loadModel` should be replaced with `mask`.
+* The mobile `AppBar` title is static and defaults to the app name. If you want to display page
+  titles, it is recommended to use the `title` prop on the `Page`.
 
 ### 🎁 New Features
 * Tabs in `TabContainerModel` now support an `icon` property on the desktop.
@@ -25,6 +30,10 @@
 * Added new method `markManaged` on `ManagedSupport`.
 * Added new function decorator `debounced`.
 * Added new function `applyMixin` providing support for structured creation of class decorators (mixins). 
+* Added `Panel` to the mobile toolkit, which offers a header element with standardized styling,
+  title, and icon, as well as support for top and bottom toolbars.
+* The mobile `AppBar` has been updated to more closely match the desktop `AppBar`, adding `icon`,
+  `leftItems`, `hideAppMenuButton` and `appMenuButtonProps` props.
 
 ## v19.0.1 - 2019-02-12